Add files via upload
Added a script that will determine if the program needs to split the dictionary again based on the number of workers.  What it does:
1) Looks in the storage location for the master dictionary and split dictionaries
2) Error checks to see if that directory exists, and if there are files there (master dictionary and split dictionaries)
3) If there are split dictionaries it uses the file name to determine the number of worker nodes previously used
4) if there are no split dictionaries, or if the directory is empty it calls the dictionary splitter script I previously wrote to create the new dictionaries.

Some of this functionality isn't in place yet, but will be once we start integrating our scripts together.  There are a few TODO items still commented into the script for later work.