Update job_manager.py
finished the section that sends work
finished the section that checks the status of the workers
finished the section that sends out the 'stop' signal to the workers
integrated the dictionarySplitter script

Need to integrage the 'removeOldDictionarires'
Need to fix a bug in the dictionarySplitter section  - split dictionary is not written to the correct path. it just copies to the same directory as the job_manager script
Needs some logic to determine when to use the dictionary splitting function.