Update worker.py
updated to put in place the start_hashcat function. Hopefully this doesn't break anything. I've tested it as a stand alone script called "run_hashcat_linux.py" that I will upload separately.

Currently the getPID functionality is commented out since I am struggling with how to implement it. os.getpid is not asychronous, meaning if I call it to get the PID of hashcat, it kills hashcat. So it needs to be called from some other function perhaps?