Update worker.py
Updated the worker.py script start_hashcat function. Changed the name to run_hashcat since it does more than just start it. Also implemented changes we discussed after class on 10/19.  It now changes the global status variable as it runs through the run_hashcat function. I've also made as many variables global as I can so they'll be easier to identify and change when we get this on AWS. I've also figured out how to get the script to wait until subprocess calls complete, so when you run this it may take a second or two for stuff to output to the shell, but it will.

I've also removed the hashtype dictionary since this and the input hash should be sanitized and checked for errors by the job manager before it is passed to the run_hashcat function.  I'll upload a standalone version as run_hashcat_linux_v4_with_global_variables.py